deprecate shared search form element; bring component into LoadedPlaylist; slight refactor to styles
