--- src/App/Playlist/LoadedPlaylist/SearchForm/Form.scss
@@ -1,115 +1,0 @@
-@import 'styles/_transitions.scss';
-@import 'styles/_colors.scss';
-
-.playlist-search-form {
-  // background: #9ba0ff;
-  // background: #8f8cee;
-  // background: #4d4d4d;
-  background: white;
-
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  .search-form {
-    width: 100%;
-
-    .search-bar-container {
-      position: relative;
-      display: flex;
-
-      .input-container {
-        width: 100%;
-        .playlist-input {
-          width: calc(100% - 92px);
-          text-overflow: ellipsis;
-
-          border: none;
-          outline: none;
-
-          padding: 24px 0px 24px 24px;
-          // width: 100%;
-
-          font-size: 16px;
-          font-family: 'Roboto Mono';
-
-          &::placeholder {
-            font-family: 'Playfair Display';
-            transition: 0.55s $transition-style;
-            padding: 8px;
-          }
-          &:focus::placeholder {
-            color: lighten(#999, 10%);
-          }
-        }
-
-        .playlist-input-label {
-          position: relative;
-          display: block;
-          margin-top: -2px;
-          margin-left: -100px; //!!!!!!!!!!!!
-          height: 5px;
-          width: 0%;
-          background-color: $barely-purple;
-          transition: 0.95s cubic-bezier(0.79, -0.01, 0.15, 0.97);
-          &.focused {
-            width: calc(100vw - 200px); //!!!!!!!!!!!!!
-          }
-          &.valid {
-            background-color: $purple;
-          }
-          &.invalid {
-            background-color: lighten(#e44e49, 10%);
-          }
-        }
-      }
-
-      .cta-container {
-        position: absolute;
-        right: 0;
-        display: flex;
-        align-items: center;
-        height: calc(100% - 5px);
-        pointer-events: none;
-        .cta {
-          transition: 0.55s $transition-style;
-          fill: white;
-          margin: 24px;
-        }
-        &.valid {
-          cursor: pointer;
-          pointer-events: all;
-          .cta {
-            fill: $purple;
-          }
-          &:hover {
-            transform: scale(1.1);
-          }
-        }
-      }
-    }
-
-    .form-footer-container {
-      // width: 100%;
-      width: calc(100vw - 200px); //!!!!!!!!!!!!!
-      margin-left: -100px; //!!!!!!!!!!!!!
-      height: 50px;
-      display: flex;
-      justify-content: center;
-      border-radius: 0px 0px 4px 4px;
-      color: white;
-      transition: 0.55s $transition-style;
-      font-family: 'Roboto Mono';
-      .footer-content {
-        padding-top: 12px;
-      }
-      &.invalid {
-        // font-size: 15px;
-        color: darken(#e44e49, 5%);
-        // background-color: lighten(#e44e49, 10%);
-      }
-    }
-  }
-}
