--- src/App/Search/Form/Form.scss
@@ -1,79 +1,0 @@
-@import 'styles/_colors.scss';
-
-.splash-search-form {
-  width: 100%;
-  .search-form {
-    margin-top: 50px;
-    width: 100%;
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    .playlist-input {
-      border: none;
-      border-bottom: 1px solid #999999;
-      outline: none;
-
-      padding-bottom: 8px;
-      width: 400px;
-
-      font-size: 16px;
-      font-family: 'Roboto Mono';
-
-      &::placeholder {
-        font-family: 'Playfair Display';
-        transition: 0.75s cubic-bezier(0.165, 0.84, 0.44, 1);
-        padding: 8px;
-      }
-      &:focus::placeholder {
-        color: lighten(#999, 10%);
-      }
-    }
-
-    .playlist-input-label {
-      position: relative;
-      display: block;
-      margin-top: -2px;
-      height: 3px;
-      width: 0%;
-      background-color: $barely-purple;
-      transition: 0.75s cubic-bezier(0.165, 0.84, 0.44, 1);
-      &.focused {
-        width: 100%;
-      }
-      &.valid {
-        background-color: $blackberry;
-      }
-      &.invalid {
-        background-color: lighten(#e44e49, 10%);
-      }
-    }
-
-    .form-footer-container {
-      width: 100%;
-      height: 50px;
-      display: flex;
-      justify-content: center;
-      border-radius: 0px 0px 4px 4px;
-      color: white;
-      transition: 0.75s cubic-bezier(0.165, 0.84, 0.44, 1);
-      font-family: 'Roboto Mono';
-      .footer-content {
-        padding-top: 12px;
-      }
-      &.valid {
-        cursor: pointer;
-        background-color: $light-purple;
-        &:hover {
-          background-color: $quite-light-purple;
-          letter-spacing: 5px;
-        }
-      }
-      &.invalid {
-        // font-size: 15px;
-        color: darken(#e44e49, 5%);
-        // background-color: lighten(#e44e49, 10%);
-      }
-    }
-  }
-}
