--- src/App/Search/Form/index.js
@@ -1,11 +1,0 @@
-import React from 'react'
-import SearchForm from 'Elements/SearchForm'
-import './Form.scss'
-
-const StyledForm = () => (
-  <div className='splash-search-form'>
-    <SearchForm ctaPosition='bottom' />
-  </div>
-)
-
-export default StyledForm
