--- src/App/Search/Search.scss
@@ -1,27 +1,0 @@
-@import 'styles/_card.scss';
-@import 'styles/_colors.scss';
-
-.search-page {
-  background: $pink;
-  width: 100vw;
-  height: 100vh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  .card-container {
-    @extend .card;
-    width: 600px;
-    height: 200px;
-    display: flex;
-    flex-direction: column;
-    // justify-content: center;
-    align-items: center;
-
-    .greeting-container {
-      margin-top: 48px;
-      font-style: italic;
-      font-size: 34px;
-      line-height: 20px;
-    }
-  }
-}
